+++ kythe/cxx/indexer/cxx/testdata/test_kindex.sh
@@ -12,7 +12,8 @@ mkdir -p "${OUT_DIR}"
     "${BASE_DIR}/kindex_test.unit" \
     "${BASE_DIR}/kindex_test.header" \
     "${BASE_DIR}/kindex_test.main"
-"${INDEXER}" "${TEST_INDEX}" > "${OUT_DIR}/kindex_test.entries"
+"${INDEXER}" "${TEST_INDEX}" --ignore_unimplemented=false \
+    > "${OUT_DIR}/kindex_test.entries"
 cat "${OUT_DIR}/kindex_test.entries" \
     | "${VERIFIER}" "${BASE_DIR}/kindex_test.verify"
 # The second test (which is useless unless the first succeeds) checks that
@@ -22,6 +23,7 @@ cat "${OUT_DIR}/kindex_test.entries" \
     "${BASE_DIR}/kindex_repo_test.header" \
     "${BASE_DIR}/kindex_repo_test.header2" \
     "${BASE_DIR}/kindex_repo_test.main"
-"${INDEXER}" "${REPO_TEST_INDEX}" > "${OUT_DIR}/kindex_repo_test.entries"
+"${INDEXER}" "${REPO_TEST_INDEX}" --ignore_unimplemented=false \
+    > "${OUT_DIR}/kindex_repo_test.entries"
 cat "${OUT_DIR}/kindex_repo_test.entries" \
     | "${VERIFIER}" "${BASE_DIR}/kindex_repo_test.verify"